Adiciona localização junto ao componente de eventos
