+++ src/app/app-routing.module.ts
@@ -9,7 +9,6 @@ import { CadastroComponent } from './auth/cadastro/cadastro.component';
 import { EventosComponent } from './eventos/eventos.component';
 import { EventosDoadorComponent } from './eventos/eventos-doador/eventos-doador.component';
 import { PerfilDoadorComponent } from './perfil-doador/perfil-doador.component';
-import { LocalizacaoComponent } from './localizacao/localizacao.component';
 
 const routes: Routes = [
   { path: '', component: HomeComponent },
@@ -19,7 +18,6 @@ const routes: Routes = [
   { path: 'eventos', component: EventosComponent },
   { path: 'eventos-doador', component: EventosDoadorComponent, canActivate: [AuthGuard] },
   { path: 'perfil-doador', component: PerfilDoadorComponent},
-  { path: 'localizacao', component: LocalizacaoComponent },
 ];
 
 @NgModule({
