+++ src/app/app.module.ts
@@ -35,7 +35,6 @@ import { CadastroComponent } from './auth/cadastro/cadastro.component';
 import { EventosComponent } from './eventos/eventos.component';
 import { EventosDoadorComponent } from './eventos/eventos-doador/eventos-doador.component';
 import { PerfilDoadorComponent } from './perfil-doador/perfil-doador.component';
-import { LocalizacaoComponent } from './localizacao/localizacao.component';
 
 @NgModule({
   declarations: [
@@ -49,7 +48,6 @@ import { LocalizacaoComponent } from './localizacao/localizacao.component';
     EventosComponent,
     EventosDoadorComponent,
     PerfilDoadorComponent,
-    LocalizacaoComponent,
   ],
   imports: [
     BrowserModule,
