+++ src/app/eventos/eventos.component.html
@@ -23,11 +23,11 @@
         <mat-select [(value)]="eventosOrder">
           <mat-option value="data_final">Tempo restante</mat-option>
           <mat-option value="nome">Nome</mat-option>
-          <!-- <mat-option value="distancia">Mais perto</mat-option> -->
+          <mat-option value="distancia">Proximidade</mat-option>
         </mat-select>
     </mat-form-field>
 </div>
-
+<output></output>
 <div *ngFor="let evento of eventos | filterBy: eventosFilter | orderBy: eventosOrder">
     <div *ngIf="mostraEvento(evento.data_final)">
         <mat-card>
