--- src/app/localizacao/localizacao.component.html
@@ -1,3 +1,0 @@
-<p>localizacao works!</p>
-
-<div id="out"></div>
